Bug/int64 for duration and throttling

+++ src/lib/ngsi/Duration.h
@@ -41,7 +41,7 @@ class Duration
 {
 public:
   std::string   string;
-  int           seconds; 
+  long long     seconds; 
 
 private:
   bool          used;
@@ -54,7 +54,7 @@ public:
   bool          isEmpty(void);
   std::string   render(Format format, std::string indent, bool comma = true);
   std::string   check(RequestType requestType, Format format, std::string indent, std::string predetectedError, int counter);
-  int           parse(void);
+  long long     parse(void);
   void          present(std::string indent);
   void          release(void);
 };
